backend/requirements.txt
sounddevice
numpy
python-dotenv
google-cloud-speech
google-cloud-texttospeech
openai
pynput
pygame 